+++ Atlas.Application/Services/IBotCallbacksService.cs
@@ -5,7 +5,7 @@ namespace Atlas.Application.Services
 {
     public interface IBotCallbacksService
     {
-        Task<string> UpdateStatusAsync(int telegramUserId, bool isDevVersionBot, Guid orderId, int status);
+        Task<string> UpdateStatusAsync(long telegramUserId, bool isDevVersionBot, Guid orderId, int status);
     }
 }
